+++ tests/tests_search/test_bayes.py
@@ -9,19 +9,11 @@ import sweeps.search.bayes as bayes
 from sweeps._types import integer, floating, ArrayLike
 from sweeps import SweepRun, RunState, next_run, SweepConfig
 
-from .test_random_search import check_that_samples_are_from_the_same_distribution
-
-
-def squiggle(x: ArrayLike) -> np.floating:
-    # the maximum of this 1d function is at x=2 and the minimum is at ~3.6 over the
-    # interval 0-5
-    return np.exp(-((x - 2) ** 2)) + np.exp(-((x - 6) ** 2) / 10) + 1 / (x**2 + 1)
-
-
-def rosenbrock(x: ArrayLike) -> np.floating:
-    # has a minimum at (1, 1, 1, 1, ...) for 4 <= ndim <= 7
-    return np.sum((x[1:] - x[:-1] ** 2.0) ** 2.0 + (1 - x[:-1]) ** 2.0)
-
+from ..utils import (
+    squiggle,
+    rosenbrock,
+    check_that_samples_are_from_the_same_distribution,
+)
 
 def run_bayes_search(
     f: Callable[[SweepRun], floating],
@@ -873,7 +865,7 @@ def test_that_constant_parameters_are_sampled_correctly():
 
 def test_metric_extremum_in_bayes_search():
     # from https://console.cloud.google.com/logs/query;query=ygnwe8ptupj33get%0A;timeRange=2021-08-03T21:34:50.082Z%2F2021-08-03T21:34:59.082Z;summaryFields=:false:32:beginning;cursorTimestamp=2021-08-03T21:34:51.189649752Z?project=wandb-production
-    data_path = f"{os.path.dirname(__file__)}/data/ygnwe8ptupj33get.decoded.json"
+    data_path = f"{os.path.dirname(__file__)}/../data/ygnwe8ptupj33get.decoded.json"
     with open(data_path, "r") as f:
         data = json.load(f)
     data["jsonPayload"]["data"]["config"]["metric"]["impute"] = "worst"
